added another status to line_item
